не вижу смысла в интерфейсе презентера
